Add TemporalBaseColumn, more extraneous methods to other files
