+++ python/cudf/cudf/core/column/string.py
@@ -17,7 +17,6 @@ import pylibcudf as plc
 
 import cudf
 import cudf.core.column.column as column
-import cudf.core.column.datetime as datetime
 from cudf.api.types import is_integer, is_scalar
 from cudf.core._internals import binaryop
 from cudf.core.buffer import Buffer, acquire_spill_lock
@@ -31,6 +30,7 @@ from cudf.utils.dtypes import (
     can_convert_to_column,
     dtype_to_pylibcudf_type,
 )
+from cudf.utils.temporal import infer_format
 from cudf.utils.utils import is_na_like
 
 if TYPE_CHECKING:
@@ -6087,7 +6087,7 @@ class StringColumn(column.ColumnBase):
             format = ""
         else:
             # infer on host from the first not na element
-            format = datetime.infer_format(not_null.element_indexing(0))
+            format = infer_format(not_null.element_indexing(0))
         return self.strptime(dtype, format)  # type: ignore[return-value]
 
     def as_timedelta_column(
