+++ python/cudf/cudf/core/tools/datetimes.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import math
 import re
 import warnings
-from typing import Literal
+from typing import TYPE_CHECKING, Literal
 
 import numpy as np
 import pandas as pd
@@ -16,11 +16,16 @@ import pylibcudf as plc
 
 import cudf
 from cudf.api.types import is_integer, is_scalar
-from cudf.core import column
 from cudf.core.buffer import acquire_spill_lock
+from cudf.core.column.column import ColumnBase, as_column
 from cudf.core.index import ensure_index
 from cudf.core.scalar import pa_scalar_to_plc_scalar
 from cudf.utils.dtypes import CUDF_STRING_DTYPE
+from cudf.utils.temporal import infer_format, unit_to_nanoseconds_conversion
+
+if TYPE_CHECKING:
+    from cudf.core.column.datetime import DatetimeColumn
+
 
 # https://github.com/pandas-dev/pandas/blob/2.2.x/pandas/core/tools/datetimes.py#L1112
 _unit_map = {
@@ -240,7 +245,7 @@ def to_datetime(
 
             times_column = None
             factor_denominator = (
-                column.datetime._unit_to_nanoseconds_conversion["s"]
+                unit_to_nanoseconds_conversion["s"]
                 if np.datetime_data(col.dtype)[0] == "s"
                 else 1
             )
@@ -261,8 +266,7 @@ def to_datetime(
                             )
 
                     factor = (
-                        column.datetime._unit_to_nanoseconds_conversion[u]
-                        / factor_denominator
+                        unit_to_nanoseconds_conversion[u] / factor_denominator
                     )
 
                     if times_column is None:
@@ -284,7 +288,7 @@ def to_datetime(
             return cudf.Series._from_column(col, index=arg.index)
         else:
             col = _process_col(
-                col=column.as_column(arg),
+                col=as_column(arg),
                 unit=unit,
                 dayfirst=dayfirst,
                 infer_datetime_format=infer_datetime_format,
@@ -326,7 +330,7 @@ def _process_col(
 ):
     if col.dtype.kind == "f":
         if unit not in (None, "ns"):
-            col = col * column.datetime._unit_to_nanoseconds_conversion[unit]
+            col = col * unit_to_nanoseconds_conversion[unit]
 
         if format is not None:
             # Converting to int because,
@@ -351,8 +355,8 @@ def _process_col(
     elif col.dtype.kind in "iu":
         if unit in ("D", "h", "m"):
             factor = (
-                column.datetime._unit_to_nanoseconds_conversion[unit]
-                / column.datetime._unit_to_nanoseconds_conversion["s"]
+                unit_to_nanoseconds_conversion[unit]
+                / unit_to_nanoseconds_conversion["s"]
             )
             col = col * factor
 
@@ -384,7 +388,7 @@ def _process_col(
                         f"{dayfirst=} not implemented "
                         f"when {format=} and {infer_datetime_format=}."
                     )
-                format = column.datetime.infer_format(
+                format = infer_format(
                     element=col.element_indexing(0),
                     dayfirst=dayfirst,
                 )
@@ -633,7 +637,7 @@ class DateOffset:
 
     def _datetime_binop(
         self, datetime_col, op, reflect=False
-    ) -> column.DatetimeColumn:
+    ) -> DatetimeColumn:
         if reflect and op == "__sub__":
             raise TypeError(
                 f"Can not subtract a {type(datetime_col).__name__}"
@@ -656,9 +660,7 @@ class DateOffset:
                             )
                         )
                 else:
-                    datetime_col += column.as_column(
-                        value, length=len(datetime_col)
-                    )
+                    datetime_col += as_column(value, length=len(datetime_col))
 
         return datetime_col
 
@@ -845,7 +847,7 @@ def date_range(
         start = dtype.type(start, unit).astype(np.dtype(np.int64))
         end = dtype.type(end, unit).astype(np.dtype(np.int64))
         arr = np.linspace(start=start, stop=end, num=periods).astype(dtype)
-        result = column.as_column(arr)
+        result = as_column(arr)
         return cudf.DatetimeIndex._from_column(result, name=name).tz_localize(
             tz
         )
@@ -933,7 +935,7 @@ def date_range(
             "months", 0
         )
         with acquire_spill_lock():
-            res = column.ColumnBase.from_pylibcudf(
+            res = ColumnBase.from_pylibcudf(
                 plc.filling.calendrical_month_sequence(
                     periods,
                     pa_scalar_to_plc_scalar(pa.scalar(start)),
@@ -954,7 +956,7 @@ def date_range(
         start = start.astype(np.dtype(np.int64))
         step = _offset_to_nanoseconds_lower_bound(offset)
         arr = range(int(start), int(stop), step)
-        res = column.as_column(arr).astype(dtype)
+        res = as_column(arr).astype(dtype)
 
     return cudf.DatetimeIndex._from_column(
         res, name=name, freq=freq
