+++ python/cudf/cudf/testing/_utils.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2020-2024, NVIDIA CORPORATION.
+# Copyright (c) 2020-2025, NVIDIA CORPORATION.
 
 import itertools
 import string
@@ -18,10 +18,10 @@ from numba.cuda.cudaimpl import lower as cuda_lower
 import pylibcudf as plc
 
 import cudf
-from cudf.core.column.timedelta import _unit_to_nanoseconds_conversion
 from cudf.core.udf.strings_lowering import cast_string_view_to_udf_string
 from cudf.core.udf.strings_typing import StringView, string_view, udf_string
 from cudf.utils import dtypes as dtypeutils
+from cudf.utils.temporal import unit_to_nanoseconds_conversion
 
 supported_numpy_dtypes = [
     "bool",
@@ -244,7 +244,7 @@ def gen_rand(dtype, size, **kwargs):
         time_unit, _ = np.datetime_data(dtype)
         high = kwargs.get(
             "high",
-            int(1e18) / _unit_to_nanoseconds_conversion[time_unit],
+            int(1e18) / unit_to_nanoseconds_conversion[time_unit],
         )
         return pd.to_datetime(
             rng.integers(low=low, high=high, size=size), unit=time_unit
