+++ python/cudf/cudf/utils/utils.py
@@ -6,7 +6,6 @@ import functools
 import os
 import traceback
 import warnings
-from typing import Any
 
 import numpy as np
 import pandas as pd
@@ -15,7 +14,6 @@ import pylibcudf as plc
 import rmm
 
 import cudf
-from cudf.core import column
 from cudf.core.buffer import as_buffer
 from cudf.utils.dtypes import SIZE_TYPE_DTYPE
 
@@ -394,60 +392,3 @@ def _is_same_name(left_name, right_name):
             return same
         except TypeError:
             return False
-
-
-def _all_bools_with_nulls(lhs, rhs, bool_fill_value):
-    # Internal utility to construct a boolean column
-    # by combining nulls from `lhs` & `rhs`.
-    if lhs.has_nulls() and rhs.has_nulls():
-        result_mask = lhs._get_mask_as_column() & rhs._get_mask_as_column()
-    elif lhs.has_nulls():
-        result_mask = lhs._get_mask_as_column()
-    elif rhs.has_nulls():
-        result_mask = rhs._get_mask_as_column()
-    else:
-        result_mask = None
-
-    result_col = column.as_column(
-        bool_fill_value, dtype=np.dtype(np.bool_), length=len(lhs)
-    )
-    if result_mask is not None:
-        result_col = result_col.set_mask(result_mask.as_mask())
-    return result_col
-
-
-def _datetime_timedelta_find_and_replace(
-    original_column: "cudf.core.column.DatetimeColumn"
-    | "cudf.core.column.TimeDeltaColumn",
-    to_replace: Any,
-    replacement: Any,
-    all_nan: bool = False,
-) -> "cudf.core.column.DatetimeColumn" | "cudf.core.column.TimeDeltaColumn":
-    """
-    This is an internal utility to find and replace values in a datetime or
-    timedelta column. It is used by the `find_and_replace` method of
-    `DatetimeColumn` and `TimeDeltaColumn`. Centralizing the code in a single
-    as opposed to duplicating it in both classes.
-    """
-    original_col_class = type(original_column)
-    if not isinstance(to_replace, original_col_class):
-        to_replace = cudf.core.column.as_column(to_replace)
-        if to_replace.can_cast_safely(original_column.dtype):
-            to_replace = to_replace.astype(original_column.dtype)
-    if not isinstance(replacement, original_col_class):
-        replacement = cudf.core.column.as_column(replacement)
-        if replacement.can_cast_safely(original_column.dtype):
-            replacement = replacement.astype(original_column.dtype)
-    if isinstance(to_replace, original_col_class):
-        to_replace = to_replace.astype(np.dtype(np.int64))
-    if isinstance(replacement, original_col_class):
-        replacement = replacement.astype(np.dtype(np.int64))
-    try:
-        result_col = (
-            original_column.astype(np.dtype(np.int64))
-            .find_and_replace(to_replace, replacement, all_nan)
-            .astype(original_column.dtype)
-        )
-    except TypeError:
-        result_col = original_column.copy(deep=True)
-    return result_col  # type: ignore
